修改角色权限控制，跳过roles拦截器，角色权限统一由自定义url拦截器处理。 增加批量新增 mybatis分页组件修改 等等。。。

+++ framework-dao/src/main/java/com/framework/dao/plugin/PageUtil.java
@@ -33,7 +33,7 @@ public class PageUtil{
 		}
 
 		RowBounds rowBounds = new RowBounds((pageIndex - 1) * pageSize, pageSize);
-		int total = baseMapper.countByExample(example);
+		Long total = baseMapper.countByExample(example);
 		List<T> pageList = null;
 		if (blob) {
 			pageList = baseMapper.selectByExampleWithBLOBsWithRowbounds(example,rowBounds);
